\name{print.all.func}
\alias{print.all.func}
\title{color pallettes}
\usage{
print.all.func
}
\description{
This package is a homemade color pallette from picture I have tken over the year. Each pallette contains the colors from a picutre of certain landscape/object, which should be obvious from the name of the pallette.

The current version has 10 pallettes and five colors in each.
}
\examples{
#get an overview of all color
print.all.func()

#col.df has all of the color

#use names(col.df) to see all of the names of the pallettes

#select a pallette with col.df$flower

#check a specific pallettes with
print.pallete.func(pallet.nm='beach')

}
